Update in endpoint for process

+++ Service/Interface/ICalculateAnalysisService.cs
@@ -7,6 +7,6 @@ namespace qualityservice.Service.Interface
 {
     public interface ICalculateAnalysisService
     {
-         Task<List<string>> Calculates(int productionOrderId, int furnaceQuantity,Analysis analysis);
+         Task<List<string>> Calculates(int productionOrderId, double furnaceQuantity,Analysis analysis);
     }
 }
